+++ example-app/build.gradle
@@ -42,6 +42,7 @@ dependencies {
     implementation project(":debug")
     implementation project(":location")
     implementation "io.rover:sdk:3.0.0-beta.1"
+    implementation project(":experiences")
 //    implementation project(":advertising")
     implementation project(":ticketmaster")
