+++ example-app/src/main/AndroidManifest.xml
@@ -87,10 +87,7 @@
         <!-- Registering RoverActivity from the Rover SDK in order to display experiences -->
         <activity
             android:name="io.rover.sdk.ui.containers.RoverActivity"
-            android:theme="@style/AppTheme.NoActionBar"
-            />
-
-
+            android:theme="@style/AppTheme.NoActionBar" />
     </application>
 
 </manifest>
