+++ experiences/build.gradle
@@ -19,14 +19,12 @@ android {
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
 
     }
+
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
-        // debug {
-        //     minifyEnabled true
-        // }
     }
 
     // Add Kotlin source directory to all source sets
@@ -45,43 +43,12 @@ android {
     }
 }
 
-androidExtensions {
-    // So we can use @Parcelize.  This does not expose risk to consumers; this is a build time-only
-    // consideration.
-    experimental = true
-}
-
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
-
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    // Until java.util.concurrent.Flow appears in Android SDK, import:
-    api 'org.reactivestreams:reactive-streams:1.0.2'
-
-    implementation "android.arch.work:work-runtime-ktx:1.0.1"
-
-    implementation 'com.android.support:customtabs:27.1.1'
-
-    testImplementation 'junit:junit:4.12'
-
-    // spek
-    testImplementation "org.spekframework.spek2:spek-dsl-jvm:$spek_version"
-    testImplementation "org.spekframework.spek2:spek-runner-junit5:$spek_version"
-
-    testImplementation 'com.natpryce:hamkrest:1.6.0.0'
-    testImplementation 'org.amshove.kluent:kluent:1.30'
-    testImplementation 'org.amshove.kluent:kluent-android:1.30'
-    testImplementation 'org.skyscreamer:jsonassert:1.5.0'
-
-    // JUnit 5 support.  Needed for Spek.
-    testImplementation("org.junit.jupiter:junit-jupiter-api:5.3.2")
-    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.3.2")
-
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
+    implementation project(':core')
+    implementation "io.rover:sdk:3.0.0-beta.1"
 }
 
 configurations {
@@ -137,7 +104,7 @@ task webDocsDokka (type: org.jetbrains.dokka.gradle.DokkaTask, overwrite: true) 
 
 ext {
     groupId = 'io.rover.campaigns' // package name of the project
-    artifactId = "core" // module name of the library
+    artifactId = "experiences" // module name of the library
     libVersion = rover_sdk_version
 }
 
@@ -185,7 +152,7 @@ publishing {
         production(MavenPublication) {
             artifact("$buildDir/outputs/aar/$artifactId-release.aar")
             groupId "io.rover.campaigns"
-
+            artifactId artifactId
             version rover_sdk_version
 
             //The publication doesn't know about our dependencies, so we have to manually add them to the pom
@@ -195,7 +162,7 @@ publishing {
                 //Iterate over the compile dependencies (we don't want the test ones), adding a <dependency> node for each
 
                 def apiAndImplementationDeps = configurations.implementation.allDependencies + configurations.compile.allDependencies
-
+                // get all the api *and* implementation dependencies
                 // get all the api *and* implementation dependencies
                 apiAndImplementationDeps.each {
                     if(it instanceof DefaultExternalModuleDependency) {
@@ -210,7 +177,10 @@ publishing {
 
                 // notice that project scope dependencies are not handled. Instead, we must add them
                 // as external dependencies manually.
-
+                def dependencyNode = dependenciesNode.appendNode('dependency')
+                dependencyNode.appendNode('groupId', 'io.rover.campaigns')
+                dependencyNode.appendNode('artifactId', 'core')
+                dependencyNode.appendNode('version', rover_sdk_version)
             }
 
             artifact sourcesJar {
