+++ experiences/src/main/kotlin/io/rover/campaigns/experiences/EventReceiver.kt
@@ -1,6 +1,7 @@
-package io.rover.campaigns.core.events
+package io.rover.campaigns.experiences
 
 import io.rover.campaigns.core.data.domain.Attributes
+import io.rover.campaigns.core.events.EventQueueServiceInterface
 import io.rover.campaigns.core.events.domain.Event
 import io.rover.campaigns.core.logging.log
 import io.rover.campaigns.core.streams.subscribe
@@ -71,7 +72,11 @@ private fun RoverEvent.BlockTapped.transformToEvent(): Event {
 }
 
 private fun RoverEvent.ExperienceDismissed.transformToEvent(): Event {
-    return Event("Experience Dismissed", mapOf("experience" to experienceAttributes(experience, campaignId)))
+    return Event("Experience Dismissed", mapOf("experience" to experienceAttributes(
+        experience,
+        campaignId
+    )
+    ))
 }
 
 private fun RoverEvent.ScreenDismissed.transformToEvent(): Event {
@@ -84,7 +89,11 @@ private fun RoverEvent.ScreenDismissed.transformToEvent(): Event {
 }
 
 private fun RoverEvent.ExperiencePresented.transformToEvent(): Event {
-    return Event("Experience Presented", mapOf("experience" to experienceAttributes(experience, campaignId)))
+    return Event("Experience Presented", mapOf("experience" to experienceAttributes(
+        experience,
+        campaignId
+    )
+    ))
 }
 
 private fun RoverEvent.ExperienceViewed.transformToEvent(): Event {
